Add the explicit version flag.
The original C daniel implementation contained a flag (0x40) that indicated
that immediately following the flags field would be an explicit format
version number.  The length would then follow that format version number.

This was designed to allow different algorithms if necessary and was never
implemented, but it was documented in the source code.  Document it in the
source code here as well, even though it's not implemented yet.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>