Note the way to generate a mask.
In order to generate a reminder, we'll need the mask based on the password.
Due to the way XOR works, we can use the generate function and pass in the
password for the mask argument, and get back the mask that we need.  Add a
test to verify that this is the case.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>