Fix CharacterSet code to work with Opal.
Opal doesn't handle \A and \z in regular expressions, so that code
always fails to match.  It also isn't aware of the str[substr] match
syntax.  Use ^ and $ instead and str.include? substr.  The former will
cause newlines to be handled differently, but it's unavoidable and
newlines are not valid in that function's inputs anyway.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>