Fix code to work in JRuby 1.7.21.
JRuby's openssl gem requires that one call cipher.final in order to get
output (which, according to the documentation, we should have been doing
anyway).  However, it also produces more output than we want, so
truncate it to an appropriate length.  Update the Opal crypto wrapper to
expect this extra method call and work appropriately.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>