Throw an InvalidArgument for bad command line options.
Since there's already a good exception class in the optparse module that
handles command line errors, use it instead of a generic
Daniel::Exception.  Note that this changes the expected exception type,
but since this change is in the main program, it's unlikely to be
problematic for others.

This removes the last instances of raising an unsubclassed
Daniel::Exception.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>