Allow passing Reminders into Reminder.parse.
If Reminder.parse is passed a reminder, return itself.  This makes it
easier for other pieces of code to always pass a reminder to
Reminder.parse, whether it's a Reminder or a String.  Adjust the one
piece of code that had to do this itself to take advantage of this.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>