Add a PRNG.
The soon-to-come PasswordSafe v3 format requires a large number of
random bytes.  While using random bytes of course makes the output
secure against attackers, it can result in unreproducible results that
are hard to unit test.  Implement a ByteGenerator class, which is an
instance of the NIST SP800-90A HMAC_DRBG.  Add unit tests for this
class based on the NIST test vectors.

Note that outside of the tests, this class will be instantiated with a
pseudorandom salt each time, which will result in good pseudorandom
output.  This mirrors the use of HMAC_DRBG in RFC 6979, where it is used
with predictable (but partially unguessable) inputs to produce secure
outputs.

The ByteGenerator class could also be used to implement future versions
of the PasswordGenerator, resulting in a simpler design that avoids the
need to use encryption algorithms (which may be restricted in some
locales).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>