Implement version 1 password generation.
Version 1 password generation takes the existing master secret and runs
it through PBKDF2 with the specified salt (if any) and iteration count
to produce a master key.  This key is then used with HKDF-Extract to
generate a 32-byte seed and a 32-byte MAC key.  The seed and an SHA-256
hash of the JSON-encoded parameters are used as the input to an HMAC
DRBG using SHA-256, and 1024-byte chunks are generated, from which the
password is produced.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>