Make PasswordGenerator able to handle any reminders.
The PasswordGenerator class was previously limited to handling one
format version at a time, which was specified in the arguments to #new.
However, this led to reminders that were of the wrong version silently
producing incorrect values.  To fix this, instantiate an object for each
version and call the correct one based on the parameters provided.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>