daniel: clean up comparisons against zero
Rubocop dislikes these, so convert them.  This has been instrumental in
indicating that the storage of the flags as an integer has been a bad
idea (at least not very Rubyish).

Convert the one instance where we're checking for a power of two into a
helper function to make it clearer what we're doing.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>