daniel-export: print newline after password prompt
The read_passphrase function in the main program (which is inherited by
daniel-export) has relied on its callers to print the required newline
when echoing has been disabled.  Without such a newline, the checksum or
existing passphrase prompts are smashed together on the same line as the
original passphrase prompt.

daniel-export did not provide such a newline, leading to an error.
Instead of relying on providing code in all callers, simply make
read_passphrase print the required newline, so the code is centralized
and all callers benefit.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>