Fix CodeParser on Opal.
In Opal, symbols are not distinct from strings: there is but one class
and all the normal string methods can be invoked on them.  Normally,
this doesn't cause a problem, but the CodeParser class had a case
statement with when blocks that contained both Symbol and String.
Because Symbol came first, all strings were treated as symbols and
therefore were not escaped.

The distinction is only necessary because CGI.unescape does not work on
symbols in MRI; our symbols never contain characters that might be
affected by escaping.  Place String before Symbol so that both MRI and
Opal work correctly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>