Switch to using YAML.safe_load
The data we're reading is a user's configuration file, and as such, it
should be trustable.  However, since we're not planning on loading any
complex types, use YAML.safe_load instead.  Add a guard for this in the
code, as this will no longer work under 1.8.7.

This also quiets a warning from RuboCop.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>