spec/main: disable invalid RuboCop warning
RuboCop warns us to use an argument to caller instead of using an empty
argument to get all frames and then selecting from the entire list.
However, what we are actually doing here is selecting the first item
from a given frame, which is fine (and perfectly performant).  Disable
this warning as it is incorrect.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>