daniel: always use 3 decimal places for estimates
With entropy estimates, it was possible for some values to end up
formatting to an absurd number of decimal places due to typical
floating-point issues, which was unsightly.  Limit formatting to three
decimal places, which provides enough information but not too much to be
useful.

Note that the difference in difficulty between a password with 1024 bits
of entropy (an absurd amount) and 1024.001 bits is less than 7 parts in
10,000, so this should be precise enough for most people's needs.
daniel cannot actually produce passwords with that much entropy anyway
due to limitations in the cryptographic algorithms used.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>